% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfout.R
\name{pdf_callback}
\alias{pdf_callback}
\title{The main PDF callback.}
\usage{
pdf_callback(device_call, args, state)
}
\arguments{
\item{device_call}{name of device call}

\item{args}{arguments to the call}

\item{state}{rdata, gc and dd}
}
\description{
The main PDF callback.
}
