% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfout.R
\name{pdf_strWidth}
\alias{pdf_strWidth}
\title{Return the width of the given string}
\usage{
pdf_strWidth(args, state)
}
\arguments{
\item{args, state}{standard pass-through from device driver}
}
\description{
Return the width of the given string
}
