% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfout.R
\name{pdfout}
\alias{pdfout}
\title{PDF device written in R. JUst sets up \code{rdevice} to call \code{pdf_callback()}}
\usage{
pdfout(filename = "pdfout.pdf", width = 7, height = 7, ...)
}
\arguments{
\item{filename}{default: "pdfout.pdf"}

\item{width, height}{size in inches. default to 7}

\item{...}{arguments passed to \code{devout::rdevice}}
}
\description{
PDF device written in R. JUst sets up \code{rdevice} to call \code{pdf_callback()}
}
